+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `FUNC.traceCall` method to contract functions
 * Add `abi.decodeData` to decode function calldata and errors from any known ABI
 * Add `address.transfer` to send ETH to an address
+* Add `json.stringify` to convert a value to a JSON string
 
 ### Other changes